requirements.txt
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
transformers==4.40.0
opencv-python==4.9.0.80
librosa==0.10.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
matplotlib==3.8.4
seaborn==0.13.2
tqdm==4.66.4
PyYAML==6.0.1
soundfile==0.12.1
Pillow==10.3.0
